+++ lib/lightning/config/bootstrap.ex
@@ -228,11 +228,11 @@ defmodule Lightning.Config.Bootstrap do
     cleanup_cron =
       if Application.get_env(:lightning, :purge_deleted_after_days) > 0,
         do: [
-          {"2 2 * * *", Lightning.Accounts, args: %{"type" => "purge_deleted"}},
-          {"3 2 * * *", Lightning.Credentials,
+          {"15 2 * * *", Lightning.Accounts, args: %{"type" => "purge_deleted"}},
+          {"30 2 * * *", Lightning.Credentials,
            args: %{"type" => "purge_deleted"}},
-          {"4 2 * * *", Lightning.Projects, args: %{"type" => "purge_deleted"}},
-          {"5 2 * * *", Lightning.WebhookAuthMethods,
+          {"45 2 * * *", Lightning.Projects, args: %{"type" => "purge_deleted"}},
+          {"0 3 * * *", Lightning.WebhookAuthMethods,
            args: %{"type" => "purge_deleted"}}
         ],
         else: []
@@ -643,10 +643,13 @@ defmodule Lightning.Config.Bootstrap do
 
     setup_storage()
 
-    entry_points = React.get_entry_points(:lightning)
+    # Commenting this out because the React modules aren't being used in prod
+    # Utils.get_env([:esbuild, :default, :args]) returns nil in prod
+    # We should have uncomment it when we have a proper fix
+    # entry_points = React.get_entry_points(:lightning)
 
-    config :esbuild, :default,
-      args: Utils.get_env([:esbuild, :default, :args]) ++ entry_points
+    # config :esbuild, :default,
+    #   args: Utils.get_env([:esbuild, :default, :args]) ++ entry_points
 
     :ok
   end